Fixed loading float values in entload command.

+++ codemp/game/g_main.c
@@ -5902,14 +5902,14 @@ void G_RunFrame( int levelTime ) {
 
 						zyk_set_entity_field(new_ent,"classname",va("%s", content));
 
-						fx = atoi(zyk_get_file_value(this_file));
-						fy = atoi(zyk_get_file_value(this_file));
-						fz = atoi(zyk_get_file_value(this_file));
+						fx = atof(zyk_get_file_value(this_file));
+						fy = atof(zyk_get_file_value(this_file));
+						fz = atof(zyk_get_file_value(this_file));
 						zyk_set_entity_field(new_ent,"origin",va("%f %f %f",fx,fy,fz));
 
-						fx = atoi(zyk_get_file_value(this_file));
-						fy = atoi(zyk_get_file_value(this_file));
-						fz = atoi(zyk_get_file_value(this_file));
+						fx = atof(zyk_get_file_value(this_file));
+						fy = atof(zyk_get_file_value(this_file));
+						fz = atof(zyk_get_file_value(this_file));
 						zyk_set_entity_field(new_ent,"angles",va("%f %f %f",fx,fy,fz));
 
 						zyk_set_entity_field(new_ent,"spawnflags",zyk_get_file_value(this_file));
@@ -5918,14 +5918,14 @@ void G_RunFrame( int levelTime ) {
 
 						zyk_set_entity_field(new_ent,"target",zyk_get_file_value(this_file));
 
-						fx = atoi(zyk_get_file_value(this_file));
-						fy = atoi(zyk_get_file_value(this_file));
-						fz = atoi(zyk_get_file_value(this_file));
+						fx = atof(zyk_get_file_value(this_file));
+						fy = atof(zyk_get_file_value(this_file));
+						fz = atof(zyk_get_file_value(this_file));
 						zyk_set_entity_field(new_ent,"mins",va("%f %f %f",fx,fy,fz));
 
-						fx = atoi(zyk_get_file_value(this_file));
-						fy = atoi(zyk_get_file_value(this_file));
-						fz = atoi(zyk_get_file_value(this_file));
+						fx = atof(zyk_get_file_value(this_file));
+						fy = atof(zyk_get_file_value(this_file));
+						fz = atof(zyk_get_file_value(this_file));
 						zyk_set_entity_field(new_ent,"maxs",va("%f %f %f",fx,fy,fz));
 
 						zyk_set_entity_field(new_ent,"wait",zyk_get_file_value(this_file));
